requirements.txt
# OS Agent dependencies
# Install: pip install -r requirements.txt
# For CUDA: CMAKE_ARGS="-DGGML_CUDA=on" pip install llama-cpp-python

llama-cpp-python>=0.3.18
pyyaml>=6.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
prompt_toolkit>=3.0.0
